Xog project 100%

+++ main/java/tdc/edu/vn/quanly_dathang_xemay/model/ChiTietDonDatHang.java
@@ -26,7 +26,7 @@ public class ChiTietDonDatHang extends DonDatHang {
         this.soLuongDatHang = soLuongDatHang;
     }
 
-    public double getDonGia() {
+    public int getDonGia() {
         return donGia;
     }
